backend/requirements.txt
Django==3.1.14
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djoser==2.1.0
drf-extra-fields==3.2.1
gunicorn==20.1.0
pillow==9.0.0
psycopg2-binary==2.9.3
python-decouple==3.5
reportlab==3.6.5